feat(user): implement and test UpdateUser function
Added UpdateUser function in controllers/user_controller.go to update a user's information by their ID.
Implemented SQL statement to update user details.
Handled decoding of JSON request body and error handling during the update process.
Created a TestUpdateUser function in tests/user_controller_test.go to test the UpdateUser functionality.
The test assumes the existence of a user with ID 1 in the database.
This commit introduces the ability to update a user's information using the /user/{id} endpoint through a PUT request. A corresponding test has been added to ensure the correctness of the implementation.